moved typedefs to external definition file
